[TASK] Ensure `push` workflow run as `workflow_run`
Due to the fact that commiting and pushing during
a github action only using the `$GITHUB_TOKEN` of
the action does not trigger `push` workflow, it's
required to add a `workflow_run` configuration to
these workflows.

To narrow that down to specified branches AND only
specified workflows, all workflow are getting the
same name as the workflow file for easier referencing.