[FEATURE] Add `findutils` package as additional dependency
The TYPO3 core testing images only adds stuffs which is
required for the core testing itself and which can be
considered as generic project and extension testing
requirement.

The `FriendsOfTYPO3/tea` extension of the TYPO3 Best
practice team uses advanced `find` options which are
not part of the base busybox provided tools used by
the alpine base image.

Albeit this repository is considered as a demonstration
how to automatically build extended images based on the
core testing images, it already contains additions used
by different projects and extension.

This change adds the requirement for the `tea` extension
to replace the core testing images there and to use the
extended `find` binary options - to fix an issue.

Additionally, all changes images are raised to the next
minor version for the rebuild.